split admin.py file
