+++ admin_panel/bd_models/admin/__init__.py
@@ -1,0 +1,15 @@
+from .ball import BallAdmin, EconomyAdmin, RegimeAdmin
+from .ball_instance import BallInstanceAdmin
+from .guild import GuildAdmin
+from .player import PlayerAdmin
+from .trade import TradeAdmin
+
+__all__ = [
+    "BallAdmin",
+    "EconomyAdmin",
+    "RegimeAdmin",
+    "BallInstanceAdmin",
+    "GuildAdmin",
+    "PlayerAdmin",
+    "TradeAdmin",
+]
